+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [5.0.11] - 2022-05-24
+### Added
+- Added INGXLogger.partialUpdateConfig that updates only some fields of the config
+
 ### Changed
 - INGXLoggerConfig.level is now optional (ERROR is the default value)